Brought back the button with confirmation, if user wants to keep item, new component: Armory
